Ensure inputs to CallGraphGenerator::analyze_localfunction are ordered
* Currently `CallGraphGenerator::do_pass` invokes `CallGraphGenerator::analyze_localfunction` with a collection that is constructed by iterating over `self.module_manager.locals`. `self.module_manager.locals` is a set, and sets are unordered in Python, so the order in which modules are processed by `analyze_localfunction` is non-deterministic. This contributes to non-determinism in call graph generation.

Since we are already ordering the files passed to the call graph generator in `_depth_first_by_directory`, we should be able to simply update `CallGraphGenerator::do_pass` to keep track of the order in which the modules are processed. This commit introduces a new local variable `modules_by_order_analyzed`: a list to which fully qualified module names are appended after they are processed. And instead of iterating over the set `self.module_manager.local`, we iterate over the ordered list `modules_by_order_analyzed` to build the collection passed to `analyze_localfunction`.